Floating point acceleration times for focuser

+++ src/telescope/focuser/Focuser.cpp
@@ -14,8 +14,8 @@ typedef struct FocuserConfiguration {
   bool present;
   uint16_t slewRateDesired;
   uint8_t  slewRateMinimum;
-  uint8_t  accelerationTime;
-  uint8_t  rapidStopTime;
+  float    accelerationTime;
+  float    rapidStopTime;
   bool powerDown;
 } FocuserConfiguration;
